Add support for 'unlimited' in crashdump params
This commit adds support for 'unlimited' value to max_used and max_size
crashdump service-parameters.

Story: 2010893
Task: 48941

Test Plan:

PASS: Add new crashdump 'max_used' service-parameter and set its value
to 'unlimited'. Verify that the service-parameter is created and
crash-dump-manager service is executed using this value.
PASS: Modify the previous service-parameter to a value different from a
human-readable format or the word 'unlimited'. Verify that an error is
displayed and the service-parameter is not updated.
PASS: Same tests for 'max_size' parameter.

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: Ib09c0e6e21d29a5af0d7c9635c8d982f91fab373