Firewall: allow IGMP queries from 0.0.0.0
IGMP queries from address 0.0.0.0 on the cluster-host and mgmt networks
are being blocked and it causes the heartbeat traffic to stop and the
controller-0 to reboot.
IGMP queries with source address 0.0.0.0 are sent when the switch
between the controllers has IGMP snooping enabled on a VLAN without an
IP address configured.

Test Plan:
    PASS: check if IGMP rules for all networks are correct
    PASS: check if the IGMP queries from 0.0.0.0 are allowed and IGMP
          reports are replied
    PASS: check if heartbeat messages are exchanged between the
          controllers

Closes-Bug: 2039881

Signed-off-by: Caio Bruchert <caio.bruchert@windriver.com>
Change-Id: Id930edf692dbd59fecef4647904abfc58e881669