Delete Kubernetes resources on application updates
Within the application update process, it is necessary to
call "kubectl delete -k <manifest_dir>" to delete the
charts that will not be present in the new version of the
application. This way we eliminate unnecessary
remnants of secrets in the N+1 application.

Test Plan:
PASS: Upload/Apply/Remove/Delete cert-manager
PASS: Upload/Apply/Remove/Delete plataform-integ-apps
PASS: Platform-integ-app update to the new version with
changed list of the charts
PASS: Secrets no longer used were deleted
PASS: If the update fails, it must remove the secrets that
are in version N+1 and should not be in version N

Closes-Bug: 2040277

Change-Id: I1c281491d30b46a7cbf53211890bb4add021dcc8
Signed-off-by: David Barbosa Bastos <david.barbosabastos@windriver.com>