Fix the condition to delete a stuck partition in the database
In the changes of [1] review, there is a condition to delete a
partition in the database that doesn't exist in the agent report
and at the same moment, there is no puppet from the
'platform::partitions::runtime' class running.

In case of a partition with the status "Creating on unlock", it
satisfies both conditions I mentioned, because the agent won't
report it, and Puppet to create a partition won't be running.
This commit changes the behavior to not delete the partition with
this status, because it will still be created during unlock.

Additionally, a failure was also identified in the check condition
when puppet is running, which was causing the partition to be
deleted incorrectly. To fix this, an in-file flag was implemented
to identify puppet manifest execution.

[1] https://review.opendev.org/c/starlingx/config/+/889090

Test-Plan:
  PASS: AIO-SX fresh install
  PASS: AIO-DX fresh install
  PASS: create/modify/delete a partition in the
        controller-0|1 followed by a reboot and check the status
        with 'system host-disk-partition-list'.
  PASS: Restart of sysinv-conductor and/or sysinv-agent services
        during puppet manifest applying.
  PASS: AIO-SX upgrade stx 7.0 to stx 8.0
  PASS: AIO-SX Backup and Restore

Closes-Bug: 2028254

Change-Id: I2024ab841ca3edbcc140de9b4ea0fbea12044791
Signed-off-by: Gabriel de Araújo Cabral <gabriel.cabral@windriver.com>
Signed-off-by: Erickson Silva <Erickson.SilvadeOliveira@windriver.com>