Get system-local-ca secret's ca.crt (DX sc upgrade)
Modified sysinv API call for certificate installation with type
openldap_ca to extract ca data included in the certificate bundle and
include it in the 'system-local-ca' ca.crt field.

This code is called during DX subcloud upgrade and is used to keep
the current 'system-local-ca' on the subcloud consistent with the
one in the SystemController. The distcloud repo counterpart will be
addressed in a future change, and will perform this call passing a
bundle with TLS cert + CA cert + TLS key from the 'system-local-ca' in
the SystemController.

This data will be used then to recreate the secret and the
ClusterIssuer in the subcloud. If a Certificate resource owns the
'system-local-ca' secret, it will be deleted, as we don't support
auto-renewals for 'system-local-ca' certificate itself, only for the
certificates issued by it.

The proper system upgrades tests will be done together when the
support for DC systems is concluded in a future task under this Story.

Test plan:
PASS: Using sysinv API, perform the call to install a certificate with
      openldap_ca type, passing an ICA data with the form: TLS cert +
      RCA cert + TLS key in a .pem file. Observe that the
      'system-local-ca' secret is recreated with the data fields:
      'ca.crt' - RCA cert (base64)
      'tls.crt' - ICA TLS cert (base64)
      'tls.key' - ICA TLS key (base64)

PASS: Using sysinv API, perform the call to install a certificate with
      openldap_ca type, passing an RCA data with the form: RCA cert +
      RCA cert + RCA key in a .pem file. Observe that the
      'system-local-ca' secret is recreated with the data fields:
      'ca.crt' - RCA cert (base64)
      'tls.crt' - RCA cert (base64)
      'tls.key' - RCA key (base64)

PASS: Using sysinv API, perform the call to install a certificate with
      openldap_ca type, passing an RCA or ICA data with the form:
      CA cert + CA key in a .pem file.  Observe that the
      'system-local-ca' secret is recreated with the data fields:
      'ca.crt' - '' (empty string)
      'tls.crt' - CA cert (base64)
      'tls.key' - CA key (base64)

Story: 2009811
Task: 48906

Change-Id: I6fb5a04567dcfc3b1b28750cac0377e16fbdf95e
Signed-off-by: Marcelo Loebens <Marcelo.DeCastroLoebens@windriver.com>