Fix unhashable type: 'slice' error in upgrade
The previous return type of the function get_upgrade_msg() caused
issues in the dcmanager orchestrator during subcloud upgrades.

With this small change, the output value from this function can now
be read directly without the need for any specific parsing.

Test Plan:
PASS: Create an upgrade strategy, force errors, and verify the accessibility of the output from the Systemcontroller.

Closes-bug: 2043408

Change-Id: I336014269593c504c566a6a902071c160696992b
Signed-off-by: fperez <fabrizio.perez@windriver.com>