Set default error message for for forbidden
The issue is that some error dialog shows
"the JSON object must be str, bytes or
bytearray, not NoneType" when non admin user
tries to modify configuration.

Functional-wise it is expected behavior to
reject configure operation because of
forbidden. But this message is not proper
for users.

This happens when it extracts the error message
from api, it is a case to be empty. The fix
is check empty message and show the default
message if it is empty.

Closes-bug: 2037320

Test Plan:
PASS: Confirm error message dialog contains
      proper text in detail.

Change-Id: I980df3356b60a59b19ec8b552f848e63dec3b621
Signed-off-by: Takamasa Takenaka <takamasa.takenaka@windriver.com>