Add docker filesystem class to docker runtime execution
To ensure that the docker-lv filesystem is present and mounted before
starting docker, docker.pp was updated to include additional
dependencies on the mount of the filesystem. This was done to avoid a
race condition where docker was started prior to creation of the
filesystem and would generate key files/directories only to have the
filesystem later mount causing those files to "disappear" from the
daemon. The end result was docker pulls failing due to missing files.

This change now include the 'platform::filesystem::docker' to provide
the Mount['docker-lv'] target now need by the docker configuration
classes that are executed at unlock and at runtime.

Test Plan:

PASS - creation and application of service parameter:
       docker/proxy/http_proxy, docker/proxy/https_proxy,
       docker/proxy/no_proxy. Observed successful runtime manifest
       execution and create of http-proxy.conf files for containerd and
       docker.

Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/900806
Change-Id: Ia742b17af7492f45b2fbe742a4887a0d2047747a
Partial-Bug: #2043399
Signed-off-by: Robert Church <robert.church@windriver.com>