Resolve kubeadm alpha Check in kube-cert-rotation Script
The 'kube-cert-rotation' script evaluates the status
of the 'kubeadm' command to determine if it is in
alpha mode. This modification addresses an issue with
the check, ensuring it conforms to the correct usage.
After kubernetes v1.21 the 'kubeadm alpha' has been
deprecated.

TEST PLAN:

PASS: Deploy the DC (with dc-libvirt) and verify
      that alarm 250.003 is not triggered.

Closes-bug: 2045271
Change-Id: I05d7aabbb3ac35b59b78e1c4f0cb070ec671b6a7
Signed-off-by: Lucas Borges <lucas.borges@windriver.com>