Update OAM firewall
In order to use the OAM network value in the firewall we are using
the destination instead of the source (as it is done for the other
platform networks) since the OAM is a special case where outside
access is possible but needs to be limited to this network.
For ICMPv6 the link-local (unicast and multicast) networks
are added.

Test Plan:
[PASS] Install an AIO-SX in IPv4 and check OAM traffic termination
[PASS] Install an AIO-DX in IPv4 and check OAM traffic termination
[PASS] Install an AIO-SX in IPv6 and check OAM traffic termination
[PASS] Install an AIO-DX in IPv6 and check OAM traffic termination

Story: 2010591
Task: 49214

Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>
Change-Id: Icededa544de12545d1cb8644b47ce941d89d5f56