Prevent mgmt gateway overwriting default route
The default route is setup using the OAM IP gateway normally.
For a standalone AIO-SX, if the management gateway is specified in
the address pool creation during management reconfiguration, a
second gateway is setup for the management interface. The ifupdown
package ensures that there is only one gateway based on the configured
gateway(s). For this reason the default route using OAM IP gateway may
get overwrittened by the management gateway.

For a DC subcloud, the management gateway is used to configure the
system controller gateway IP for communication between the subcloud and
system controller. Therefore, it is required to configure the managemet
gateway for a subcloud.

The fix is to add a semantic check to prevent the addition of the
management gateway for a standalone AIO-SX.

Test Plan:
PASS: AIO-SX management reconfiguration
PASS: AIO-SX subcloud management reconfiguration

Story: 2010722
Task: 49218

Signed-off-by: Teresa Ho <teresa.ho@windriver.com>
Change-Id: Ib50f3942fc7a81acf4db1dfa7fbf86226e5e245a