Revert "Semantic check for proper sriov plugin configuration"
This reverts commit 16e800bcf9a9fc2bd97cc05ba0213091352c6130.

Reason for revert:
This validation is preventing a 2 step approach to the configuration,
first configuring labels and later adding the interface configuration.
With the change:
https://review.opendev.org/c/starlingx/stx-puppet/+/869852
there is no reason to keep this validation as the pod sriov device
plugin will no longer generate log spam since a dummy config file
will be created, avoiding the state of CrashBackLoop.

Closes-Bug: 2002447

Change-Id: Ice4def2a23c5fe5fc0311bafff1a8ea5051e1609
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>