Moving apparmor_config_status to latest migration script
A new column apparmor_config_status was added to the i_host table
to track the runtime apparmor config application status. The
column creation was done in existing script 126 which is causing
the issue in upgrade migration. Moved this column creation to the
new migration script.

Closes-Bug: 2045965

Test plan:
PASS: AIO-SX: Install AIO-SX and verify if apparmor_config_status
      column is created in i_host table

Change-Id: I34d617ef084dacbad0b7153bd186a00e9482b38b
Signed-off-by: Jagatguru Prasad Mishra <jagatguruprasad.mishra@windriver.com>