Change host config_target when trying to reapply
Commit [1] introduced the capability for retrying runtime
manifests that have not reported success back to the conductor
after a specified timeout. During configuration-intensive
periods like host provisioning, the host config_target can
change multiple times due to various manifests being sent
to it, so in a scenario where a manifest is lost, when it
is retried, the host will apply it and then remain out-of-date
because the config_target wasn't updated accordingly.

This commit updates the host config_target before the system
attempts to reapply a runtime manifest whose config_uuid
differs from the current host config_target.

This commit also changes the message type of runtime_config
duplicate to "info", since it is expected to already exist
in the database in the reapply scenario and is harmless to
the system.

[1] https://review.opendev.org/c/starlingx/config/+/894544

Test Plan
PASS: install/bootstrap/unlock single and multinode deployments
PASS: force the issue to be reproduced for a host and verify:
      - Manifest is reapplied on the host by the conductor
      - The configuration is really applied to the host
        (e.g. in this case filesystem "docker" was increased)
      - Host config_target field is updated on the database
      - No config out-of-date alarms are present
PASS: execute common configuration operations and verify that
      behavior is unchanged
PASS: run intensive runtime configuration stress test and verify
      that the host is configured and no config out-of-date alarms
      are present on the system

Closes-bug: 2045704

Change-Id: Id2e41f990bebedd443e8eca76fc05a2d4b910aaf
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>