Set the alarm to update the reason text
On alarm audit, we are clearing and re-raising the
same alarm to update the reason text, this is overwhelming
the system with alarms on every audit. This change addresses
this issue by doing a set request to update the alarm instead
of clear,set and also only raises the alarm if the reason text
is changed.

Test Cases:
PASS: Create a certificate that expires in 30days, run
      full audit to raise the alarm and verify that on
      active alarm audit, system is replacing the exiting
      alarm with new alarm and corresponding event is raised
      when the reason text is changed.
PASS: Create a certificate that expires with in 30days,
      run the full cert-alarm audit to raise the alarm.
      now run the hourly audit and verify that the system
      is not raising an alarm if the alarm reason text is
      not changed.
PASS: Create multiple certificates, one that changes reason
      text periodically and the other never changes its
      reason text, monitor active alarm audit and verify
      that set alarm request is raised only on the
      certificate where the reason text is changed and
      for the other certificate no alarm is raised.
PASS: Create a certificate that expires with in 30days,
      run the full cert-alarm audit to raise the alarm.
      now adjust duration >30days and renewbefore to 15days
      of the certificate, observe on active alarm audit
      the alarm is cleared and a corresponding clear event
      is raised.
PASS: Build the ISO with changes and install the lab,run all
      the above testcases have successfully passed

Closes bug: 2046374

Change-Id: Ia908bc5331b6a2af97d4995017f0b99293093094
Signed-off-by: amantri <ayyappa.mantri@windriver.com>