Armada image not found by script 76
This change fixes the removal of Armada images from registry.
Initially it was thought that it would only have the
airshipt/armada image, but there is also starlingx/armada-image.
To resolve this, the starlingx/armada-image key was added to
the images to be deleted.

Test Plan:
PASS Build pkgs and build image
PASS Upgrade SX stx-8 -> master
PASS When the 76-remove-armada-if-unused.py script is run
no error appears and at the end of activation no Armada
image is found.

Closes-Bug: 2048400

Change-Id: I91ed6a2f8bd0478d845a68716ea6e19b978c47f9
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>