Enable SR-IOV for slave interface of bond
The number of VFs was not being populated in the
/sys/class/net/interface file by the current implementation.
The member interface of the bond should be of the SR-IOV
class in the current implementation.
In this commit, the sriov_numvfs for the slave interface of
the AE bond has been updated.

Test plan:
PASSED: Build process is successful.
PASSED: - AE active-standby bond creation of pci-sriov interface
          AIO-SX (CLI,Horizon,DM).
        - Verified AIO-SX vf number through
          cat /sys/bus/pci/devices/0000:51:00.1/sriov_numvfs.
        - Verified VF created through lspci.
        - Assigned mgmt and cluster-host network to bond.
        - Verified system host lock/unlock.
        - Verified on IPV4 and IPV6.

Story: 2010706
Task: 49398

Change-Id: If293980d81ffa3fa67246fcfa45d795a733914d1
Signed-off-by: Mayank Patel <mayank.patel@windriver.com>