Initial implementation of IPsec Auth Client
This commit adds the initial implementation for IPsec Auth Client,
responsible for connecting with IPsec Auth Server and initialize
IPsec PKI Auth procedure. The IPsec Auth Client expects a certificate
at the end of this procedure to configure swanctl and enable IPsec with
other hosts in MGMT network environment.

The IPsec Auth Client should be able to initialize IPsec Auth operation
by firstly connect and request IPsec Auth Server to generate and pass
keys and certificates through the execution of the procedure. The IPsec
Auth Client must be able to create and store keys and certificates, such
as use those files and host server info to configure swanctl.

The main goal of this commit is to create an TCP client able to connect
with IPsec Auth Server, exchange, encrypt and decrypt messages with the
server by a single TCP connection in a local network environment. At
the end of its execution, IPsec must restart ipsec.service after
configure swanctl and place correctly the received and generated keys &
certificates.

Test Plan:
PASS: Build, install and bootstrap an AIO-DX system with a worker node
      associated.
PASS: In a DX system with a worker node associated, login to a
      controller or worker node and execute "ipsec-client -h" command.
      Observe that a help message is displayed in terminal screen
      specifying the command description, arguments and default values
      that may be passed in to the command line.
      All systems in this environment are in enable available active
      statuses.
PASS: In a DX system with a worker node associated, login to the
      worker node and execute "sudo ipsec-client pxecontroller"
      command. Observe that IPsec Auth client tries to connect with
      the server named as pxecontroller and perform initial-auth
      operation. The message exchange, certs, keys and configuration
      of swanctl should be printed on the screen. If any error occurr,
      the connection is closed and an error message is printed out.
      DX systems in this environment are in enable available active
      statuses. Worker node should not be in inventoried state yet.
      Perform this test with ipsec.service in inactive status.
PASS: In a DX system with a worker node associated, login to the worker
      node and execute "sudo ipsec-client pxecontroller" command
      sequential times to simulate multiple client requests targeted
      to a single IPsec server. Observe that IPsec Auth client tries to
      connect with the server named as pxecontroller and perform
      initial-auth operation. The message exchange, certs, keys and
      configuration of swanctl are printed on screen and IPsec PKI Auth
      procedure is executed for each one of the IPsec client requests.
      If any error occur, the connection is closed. DX system is in
      in enable available active status. Worker node should not be in
      inventoried state. Perform this test with ipsec.service in
      inactive status.

Story: 2010940
Task: 49418

Depends-on: https://review.opendev.org/c/starlingx/config/+/905998

Co-Authored-By: Andy Ning <andy.ning@windriver.com>
Co-Authored-By: Leonardo Mendes <Leonardo.MendesSantana@windriver.com>
Signed-off-by: Manoel Benedito Neto <manoel.beneditoneto@windriver.com>
Change-Id: I83f069b4a714b3cd7a46cecd7de81733fa69bd55