Add default ha_priority value and remove value pruning
When a user sets the ha_enabled=1 parameter, have phc2sys provide a sane
default ha_priority value for configured interfaces if one has not been
supplied. The default value used is 0, which is the lowest priority
available.

This change also removes the pruning behaviour where some parameters are
removed from the config file when ha_enabled is not set. This behaviour
is unnecessary and causes challenges for troubleshooting as it removes
parameters that the user explicitly configured.

Test Plan:
Pass: Configured phc2sys with ha_enabled set
Pass: Ensure that default ha_priority is set when not supplied by user
Pass: Ensure that user supplied ha_priority overrides default values
Pass: Ensure that ha_ parameters are not pruned when ha_enabled is not
present/set to 0

Story: 2010723
Task: 49445

Change-Id: I3ca78def65c3f2d31224c90ed1f7c997a03821fc
Signed-off-by: Cole Walker <cole.walker@windriver.com>