Use mgmt_ipsec in sysinv for ipsec request check
Currently ipsec server use inv_state in sysinv to check the validation
of the auth request from client. This is found to be problematic. This
change updated ipsec server to use "mgmt_ipsec:enabling|enabled" flag
in capabilities of the i_host table for the validation checking.

Also a minor refactoring to move the flag setting into a function in
the utils module, since the flag setting will be eventually used in
multiple places.

Test Plan:
PASS: DX deployment in VBox. Verify controller-0 and controller-1
      are installed, bootstrap and unlocked successfully with IPSec
      configured and enabled.

Story: 2010940
Task: 49558

Change-Id: I397ea29d73ad8a3b8b8ce5500a4501c7bc2fbfbc
Signed-off-by: Andy Ning <andy.ning@windriver.com>