Retrieve only the deploy plug-in helm-chart
Fetching the deploy plug-in helm-chart by using
the command "grep "deployment-manager.tgz" in the
directory /usr/local/share/applications/helm/ for
the purpose of deploying an upgrade to deploy plug-in.

Due to recent developments in the deploy plug-in
system-application, a new helm-chart is generated with
the naming convention deployment-manager_*tgz. The
current retrieval process is obtaining both
deploy plug-in and system application helm-charts.

To address this, a hyphen ("-") has been added to the
grep command ("-deployment-manager") to specifically
check for deploy plug-in and exclude the system-application.

Test-Plan:
PASS: Verify the deploy plug-in helm-chart is retrieved.
PASS: Verify the deploy plug-in is refreshed when it
      is installed

Closes-Bug: 2055699

Change-Id: I9d8ae5d491f6a9cacca1e68574c4cb4431264342
Signed-off-by: sharang bhardwaj <sharang.bhardwaj@windriver.com>