Add CONF option to set default auto_update value
Add a configuration option to sysinv.conf that sets the default behavior
for automatically updating apps if not specified in the application
metadata.

The new option is named 'missing_auto_update' and it is available under
the 'app_framework' section in sysinv.conf. It is set to 'False' in de
code base to facilitate porting this change to previous releases, which
had the auto_update option disabled by default. A separate stx-puppet
change was done to set it to 'True' by default for the next releases:
https://review.opendev.org/c/starlingx/stx-puppet/+/911384

Test Plan:
PASS: AIO-SX fresh install.
PASS: Build platform-integ-apps without the 'upgrades' metadata
      section.
      Confirm that the default value was correctly inserted into the
      'auto_update' field of the 'kube_app_bundle' database table after
      a fresh install.
PASS: Change the default 'missing_auto_update' value in sysinv.conf.
      Add a new version of platform-integ-apps without the 'upgrades'
      metadata section to /usr/local/share/applications/helm/.
      Restart sysinv-conductor.
      Confirm that the 'auto_update' column for the new version matches
      the new 'missing_auto_update' value.
      Apply platform-integ-apps and confirm that the framework complies
      with the new 'missing_auto_update' value.

Story: 2010929
Task: 49670

Change-Id: I068310bed70a5e1cf7e8668b2e673751276ed3a8
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>